+++ src/main/kotlin/com/ssak3/timeattack/common/security/JwtAuthenticationFilter.kt
@@ -2,8 +2,8 @@ package com.ssak3.timeattack.common.security
 
 import com.ssak3.timeattack.global.exception.ApplicationException
 import com.ssak3.timeattack.global.exception.ApplicationExceptionType.JWT_NOT_FOUND
-import com.ssak3.timeattack.member.infrastructure.MemberRepository
-import com.ssak3.timeattack.member.infrastructure.findByIdOrThrow
+import com.ssak3.timeattack.member.repository.MemberRepository
+import com.ssak3.timeattack.member.repository.findByIdOrThrow
 import jakarta.servlet.FilterChain
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
